1.添加TouchMouseAdaptor用于同时支持鼠标和触屏输入监听 2.现在输入监听器、输入事件均被抽象为了接口，可以自行扩展和添加至InputManager中 3.SelectConverter现在不再使用hashcode而是使用event的产生时间作为是否重新计算选中组件的判据 4.InputManager现在会自动关掉jme3默认启动的触屏鼠标模拟

+++ src/main/java/com/jarlure/ui/effect/DragEffect.java
@@ -11,7 +11,7 @@ public interface DragEffect extends AnimEffect {
      * @param x 拖拽开始时鼠标的水平坐标x值
      * @param y 拖拽开始时鼠标的垂直坐标y值
      */
-    void start(int x, int y);
+    void start(float x, float y);
 
     /**
      * 更新拖拽。对应鼠标拖拽事件
@@ -19,7 +19,7 @@ public interface DragEffect extends AnimEffect {
      * @param x 拖拽过程中鼠标的水平坐标x值
      * @param y 拖拽过程中鼠标的垂直坐标y值
      */
-    void update(int x, int y);
+    void update(float x, float y);
 
     /**
      * 结束拖拽。对应鼠标释放按键事件
